# designer-portfolio/server/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FolioStack API",
    "description": "API documentation for FolioStack - Portfolio Management Platform",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002/api",
      "description": "Development server"
    },
    {
      "url": "https://api.foliostack.app/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/works": {
      "get": {
        "tags": ["Works"],
        "summary": "Get all works",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "published", "archived"] } },
          { "name": "category", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Works retrieved successfully" }
        }
      },
      "post": {
        "tags": ["Works"],
        "summary": "Create a new work",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "url": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "category": { "type": "string" },
                  "status": { "type": "string", "enum": ["draft", "published"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Work created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/deploy": {
      "get": {
        "tags": ["Deploy"],
        "summary": "Get all deployments",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deployments retrieved successfully" }
        }
      },
      "post": {
        "tags": ["Deploy"],
        "summary": "Create a new deployment",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["projectName"],
                "properties": {
                  "projectName": { "type": "string" },
                  "customDomain": { "type": "string" },
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Deployment created successfully" },
          "400": { "description": "Validation error" },
          "409": { "description": "Deployment name already exists" }
        }
      }
    },
...
    "/ai/settings": {
      "get": {
        "tags": ["AI"],
        "summary": "Get AI settings",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "AI settings retrieved successfully" }
        }
      },
      "patch": {
        "tags": ["AI"],
        "summary": "Update AI settings",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": { "type": "string" },
                  "temperature": { "type": "number" },
                  "maxTokens": { "type": "integer" },
                  "style": { "type": "string" },
                  "apiKey": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "AI settings updated successfully" }
        }
      }
    },
...
    "/settings": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user settings",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Settings retrieved successfully" }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update user settings",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": { "type": "string", "enum": ["dark", "light"] },
                  "language": { "type": "string" },
                  "notifications": { "type": "object" },
                  "privacy": { "type": "object" },
                  "preferences": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Settings updated successfully" }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}